changed session variable to state
