+++ src/app/layout.tsx
@@ -1,17 +1,7 @@
 // src/app/layout.tsx
-import { auth } from "@/auth"
-import Link from "next/link"
-import "../app/globals.css"
-
-import { SignOut } from "./components/auth/signout-button"
-
-function NavLink({ href, children }: { href: string; children: React.ReactNode }) {
-  return (
-    <Link href={href} className="px-3 py-1 hover:text-blue-500">
-      {children}
-    </Link>
-  )
-}
+import "../app/globals.css";
+import { Navbar } from "@/app/components/Navbar";
+import { SessionProvider } from "next-auth/react"
 
 export const metadata = {
   title: "My NextAuth App",
@@ -19,30 +9,14 @@ export const metadata = {
 }
 
 export default async function RootLayout({ children }: { children: React.ReactNode }) {
-  // Check session server-side
-  const session = await auth()
 
   return (
     <html lang="en">
       <body>
-        <nav className="flex items-center justify-between px-6 py-3 border-b">
-          <div className="flex space-x-4">
-            <NavLink href="/">Home</NavLink>
-            <NavLink href="/dashboard">Dashboard</NavLink>
-          </div>
-
-          <div className="flex space-x-4">
-            {session ? (
-              // If logged in, show sign-out
-              <SignOut />
-            ) : (
-              // If not logged in, just link to the Sign In page
-              <NavLink href="/signin">Sign In</NavLink>
-            )}
-          </div>
-        </nav>
-
-        <main className="p-4">{children}</main>
+      <SessionProvider>
+          <Navbar />
+          <main>{children}</main>
+        </SessionProvider>
       </body>
     </html>
   )
